feat(ladder): rename package util -> generator https://github.com/next-step/java-ladder/pull/2264#discussion_r1835172755
