+++ src/main/java/ladder/domain/generator/RandomCrossGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain.util;
+package ladder.domain.generator;
 
 import java.util.Random;
