+++ src/test/java/ladder/domain/generator/CrossGeneratorTest.java
@@ -1,4 +1,4 @@
-package ladder.domain.util;
+package ladder.domain.generator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
